requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
idna==3.4
lxml==4.9.3
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.4.1
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
